refactor ModelStub and add AddGroupCommandTest
