+++ src/test/java/seedu/address/logic/commands/AddGroupCommandTest.java
@@ -1,0 +1,22 @@
+package seedu.address.logic.commands;
+
+//import static java.util.Objects.requireNonNull;
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertFalse;
+//import static org.junit.jupiter.api.Assertions.assertTrue;
+import static seedu.address.testutil.Assert.assertThrows;
+
+import org.junit.jupiter.api.Test;
+
+public class AddGroupCommandTest {
+
+    @Test
+    public void constructor_nullGroup_throwsNullGroupPointer() {
+        assertThrows(NullPointerException.class, () -> new AddGroupCommand(null));
+    }
+
+    @Test
+    public void execute_groupAcceptedByModul_addSuccessful() throws Exception {
+
+    }
+}
